+++ packages/amount-input/src/utils/index.ts
@@ -8,8 +8,6 @@ import { formatAmount, getCurrencySymbol } from '@alfalab/utils';
 
 import { type AmountInputProps } from '../Component';
 
-export const SEP = ',';
-
 /**
  * Форматирует введенное значение
  * @param enteredValue Значение введенное в инпут
@@ -17,22 +15,20 @@ export const SEP = ',';
  * @param minority количество минорных единиц
  */
 export function getFormattedValue(enteredValue: string, currency: CurrencyCodes, minority: number) {
-    if (enteredValue === '' || enteredValue === '-') {
+    if (!enteredValue || enteredValue === '-') {
         return enteredValue;
     }
 
     // eslint-disable-next-line prefer-const
-    let [head, tail] = enteredValue.split(SEP);
+    let [head, tail] = enteredValue.split(',');
 
     // При вводе "-," указываем, что имеется в виду "-0,"
     if (head === '-') {
         head = '-0';
-    } else if (head === '') {
-        head = '0';
     }
 
     let { majorPart } = formatAmount({
-        value: parseInt(head, 10) * minority,
+        value: Number(head) * minority,
         currency,
         minority,
         negativeSymbol: 'hyphen-minus',
@@ -43,12 +39,12 @@ export function getFormattedValue(enteredValue: string, currency: CurrencyCodes,
         majorPart = `-${majorPart}`;
     }
 
-    if (!tail && enteredValue.includes(SEP)) {
-        return majorPart.concat(SEP);
+    if (!tail && enteredValue.includes(',')) {
+        return majorPart.concat(',');
     }
 
     if (tail) {
-        return majorPart.concat(SEP, tail.slice(0, minority.toString().length - 1));
+        return majorPart.concat(',', tail.slice(0, minority.toString().length - 1));
     }
 
     return majorPart;
@@ -59,13 +55,7 @@ export function getAmountValueFromStr(str: string, minority: number) {
         return null;
     }
 
-    const numberOrNaN = Number(str.replace(SEP, '.').replace(/[^0-9.-]/g, ''));
-
-    if (Number.isNaN(numberOrNaN)) {
-        return null;
-    }
-
-    return Math.round(numberOrNaN * minority);
+    return Math.round(Number(str.replace(',', '.').replace(/[^0-9.-]/g, '')) * minority);
 }
 
 export function getCurrencyCodeWithFormat(