+++ packages/bottom-sheet/src/components/header/Component.tsx
@@ -7,22 +7,34 @@ import {
     type NavigationBarPrivateProps,
 } from '@alfalab/core-components-navigation-bar-private';
 
+import { type ColorType } from '../../types';
+
+import defaultColors from './default.module.css';
 import styles from './index.module.css';
+import invertedColors from './inverted.module.css';
 
 export type HeaderProps = Omit<NavigationBarPrivateProps, 'view' | 'size'> & {
     headerRef: RefObject<HTMLDivElement>;
     headerOffset: number;
+    colors?: ColorType;
     showSwipeMarker?: boolean;
 };
 
+const colorStyles = {
+    default: defaultColors,
+    inverted: invertedColors,
+} as const;
+
 export const Header: FC<HeaderProps> = ({
     className,
+    colors = 'default',
     sticky,
     headerRef,
     headerOffset,
     showSwipeMarker,
     title,
     children,
+    contentClassName,
     ...restProps
 }) => {
     const { setHeaderOffset, setHasHeader, headerHighlighted, onClose } =
@@ -36,6 +48,7 @@ export const Header: FC<HeaderProps> = ({
         setHasHeader(true);
     }, [setHasHeader]);
 
+    const colorStyle = colorStyles[colors];
     const hasContent = Boolean(title || children);
 
     return (
@@ -49,11 +62,14 @@ export const Header: FC<HeaderProps> = ({
             className={cn(className, {
                 [styles.headerWrapper]: showSwipeMarker,
                 [styles.headerWrapperWithoutSwipeMarker]: !showSwipeMarker,
-                [styles.highlighted]: hasContent && headerHighlighted && sticky,
+                [colorStyle.highlighted]: hasContent && headerHighlighted && sticky,
+                [colorStyle.hasContent]: hasContent,
                 [styles.sticky]: sticky,
-                [styles.hasContent]: hasContent,
             })}
-            contentClassName={cn(styles.title)}
+            contentClassName={cn(styles.title, contentClassName)}
+            titleClassName={cn(colorStyle.title)}
+            subtitleClassName={cn(styles.subtitle, colorStyle.subtitle)}
+            colors={colors}
         >
             {children}
         </NavigationBarPrivate>