+++ packages/file-upload-item/src/Component.tsx
@@ -37,6 +37,7 @@ export const FileUploadItemComponent: React.FC<FileUploadItemProps> = ({
     customContent,
     truncate,
     imageUrl,
+    backgroundColor,
 }) => {
     const [actionsPresent, setActionsPresent] = useState(false);
 
@@ -76,6 +77,7 @@ export const FileUploadItemComponent: React.FC<FileUploadItemProps> = ({
                     customContent,
                     truncate,
                     imageUrl,
+                    backgroundColor,
                     actionsPresent,
                     setActionsPresent,
                 }}