+++ packages/file-upload-item/src/context/file-upload-item-context.ts
@@ -5,6 +5,8 @@ import {
     type MouseEvent,
 } from 'react';
 
+import { type SuperEllipseProps } from '@alfalab/core-components-icon-view/super-ellipse';
+
 import { type FileUploadItemStatus } from '../types/status';
 
 type TFileUploadItemContext = {
@@ -32,6 +34,7 @@ type TFileUploadItemContext = {
     customContent?: ElementType;
     truncate?: boolean;
     imageUrl?: string;
+    backgroundColor?: SuperEllipseProps['backgroundColor'];
     actionsPresent?: boolean;
     setActionsPresent?: (present: boolean) => void;
 };
@@ -61,6 +64,7 @@ export const FileUploadItemContext = createContext<TFileUploadItemContext>({
     customContent: undefined,
     truncate: false,
     imageUrl: undefined,
+    backgroundColor: undefined,
     actionsPresent: false,
     setActionsPresent: undefined,
 });