+++ packages/international-phone-input/src/components/base-international-phone-input/Component.tsx
@@ -2,6 +2,7 @@ import React, {
     type ChangeEvent,
     forwardRef,
     type MouseEvent,
+    type RefAttributes,
     useEffect,
     useMemo,
     useRef,
@@ -11,10 +12,10 @@ import mergeRefs from 'react-merge-refs';
 import { maskitoTransform } from '@maskito/core';
 import { useMaskito } from '@maskito/react';
 
+import { type InputProps } from '@alfalab/core-components-input';
 import { type InputAutocompleteProps } from '@alfalab/core-components-input-autocomplete';
 import { type AnyObject, BaseOption } from '@alfalab/core-components-select/shared';
 import { type BaseSelectChangePayload } from '@alfalab/core-components-select/typings';
-import { isNullable } from '@alfalab/core-components-shared';
 
 import { type BaseInternationalPhoneInputProps, type Country } from '../../types';
 import {
@@ -56,12 +57,11 @@ export const BaseInternationalPhoneInput = forwardRef<
             open: openProps,
             defaultOpen,
             customCountriesList,
-            autoFill = true,
+            autoFill,
             ...restProps
         },
         ref,
     ) => {
-        const lastCountryRef = useRef<Country | null>(null);
         const countriesData = useMemo(
             () => initCountries(countries, customCountriesList),
             [countries, customCountriesList],
@@ -73,6 +73,7 @@ export const BaseInternationalPhoneInput = forwardRef<
         const [openCountry, setOpenCountry] = useState<boolean | undefined>(
             countrySelectProps?.defaultOpen,
         );
+        const beforeAutofillValueRef = useRef('');
         const [selectedCountry, setSelectedCountry] = useState<Country | undefined>(() =>
             findCountry(countriesData, value, defaultIso2, countryProp),
         );
@@ -92,7 +93,7 @@ export const BaseInternationalPhoneInput = forwardRef<
                     country,
                     clearableCountryCode,
                     preserveCountryCode,
-                    lastCountryRef,
+                    beforeAutofillValueRef,
                 ),
             [country, clearableCountryCode, preserveCountryCode],
         );
@@ -206,24 +207,16 @@ export const BaseInternationalPhoneInput = forwardRef<
             />
         );
 
-        const inputProps = {
+        const inputProps: InputProps & RefAttributes<HTMLInputElement> = {
             className: styles.component,
             ref: mergeRefs([maskRef, ref, inputRef]),
             wrapperRef: inputWrapperRef,
             addonsClassName: styles.addons,
             type: 'tel',
-            autocomplete: autoFill ? 'tel' : undefined,
+            autoComplete: 'tel',
             clear: getClear(clearProp, clearableCountryCode, value, country?.countryCode),
             ...restProps.inputProps,
-        } as const;
-
-        useEffect(() => {
-            if (!preserveCountryCode || isNullable(country)) {
-                return;
-            }
-
-            lastCountryRef.current = country;
-        }, [country, preserveCountryCode]);
+        };
 
         return Array.isArray(options) ? (
             <InputAutocomplete