+++ packages/list-header/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-list-header
 
+## 4.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-typography@5.0.3
+
 ## 4.0.2
 
 ### Patch Changes