+++ packages/password-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @alfalab/core-components-password-input
 
+## 6.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@16.0.3
+    - @alfalab/core-components-icon-button@7.0.2
+
 ## 6.0.2
 
 ### Patch Changes