+++ packages/phone-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-phone-input
 
+## 8.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-masked-input@7.0.3
+
 ## 8.0.2
 
 ### Patch Changes