+++ packages/tab-bar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @alfalab/core-components-tab-bar
 
+## 3.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@1.1.1
+    - @alfalab/core-components-typography@5.0.3
+
 ## 3.0.2
 
 ### Patch Changes